feat: Adicionando gamificação ao projeto.

+++ teu_futuro/schemas/atividade.py
@@ -5,7 +5,7 @@ class AtividadeSchema(Schema):
     id = fields.Int(allow_none=False)
     nome = fields.Str(allow_none=False)
     descricao = fields.Str(allow_none=False)
-    peso = fields.Integer(allow_none=False)
+    peso = fields.Decimal(allow_none=False)
 
 
 class ResolucaoAtividadeSchema(Schema):
